+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - **buession-canal-core：** 修复 buession-beans bean 转换导致的数据丢失的 BUG
 - **buession-canal-core：** 修复 @CanalEventListener 仅指定 schema 或 table 服务映射方法的 BUG
+- **buession-canal-core：** 修复无法获取 @CanalEventListener schema 和 table 的 BUG
 
 
 ---